requirements.txt
# Core Slack dependencies
slack-bolt>=1.19.0
slack-sdk>=3.26.0

# OpenAI for LLM orchestration
openai>=1.51.0

# HTTP client for API calls
httpx>=0.27.0

# Supabase client
supabase>=2.3.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and utilities
structlog>=23.2.0
rich>=13.7.0

# Testing (optional)
pytest>=7.4.4
pytest-asyncio>=0.23.2
